group apis corrected

+++ chatAppServer/app/Http/Controllers/GroupController.php
@@ -9,15 +9,16 @@ use DB;
 
 class GroupController extends Controller
 {
+
     public function create(Request $request){
         $group = new Group;
         $group->name = $request->name;
-        $group->creator_id = $request->creator_id;
+        $group->creator_id = auth()->user()->id;
         $group->is_channel = $request->is_channel;
 
         $group->save();
         $group->members()->sync([
-            $request->creator_id
+            auth()->user()->id
         ]);
 
         return response()->json([
@@ -27,6 +28,10 @@ class GroupController extends Controller
     }
     public function delete(Request $request){
         $group = Group::find($request->group_id);
+        if(is_null($group))
+            return response()->json(['message' => 'not found',
+                ],200);
+        
         $group->delete();
 
         return response()->json([
@@ -44,16 +49,53 @@ class GroupController extends Controller
         ],200);
     }
     public function members(Request $request){
+
         $group = Group::find($request->group_id);
+        if(is_null($group))
+            return response()->json(['message' => 'not found',
+                ],200);
+        if(! $group->members->contains(auth()->user()))
+            return response()->json(['message' => 'not allowed',
+                ],403);
         
+        return response()->json(['group' => $group->members,
+            ],200);
+    }
+    public function deleteMember(Request $request){
+        $group = Group::find($request->group_id);
+        $user = User::find($request->user_id);
+        
+        if(is_null($group) or is_null($user) )
+            return response()->json(['message' => 'group or user not found',
+                ],200);
+                
+        if($group->creator_id != auth()->user()->id)
+            return response()->json(['message' => 'not allowed',
+                ],403);
+
+        DB::table('group_user')->where('user_id', $user->id)->where('group_id', $group->id)->delete();
+
         return response()->json([
-            'members' => $group ->members
-        ],200);
+            'status' => 'success',
+            'message' => 'member deleted'
+        ],200);  
     }
     public function addMember(Request $request){
         $group = Group::find($request->group_id);
-
         $user = User::find($request->user_id);
+        
+        if(is_null($group) or is_null($user) )
+            return response()->json(['message' => 'group or user not found',
+                ],200);
+                
+        if($group->creator_id != auth()->user()->id)
+            return response()->json(['message' => 'not allowed',
+                ],403);
+
+        if($group->members->contains($user))
+            return response()->json(['message' => 'user is already a member',
+                ],200);
+        
         $group->members()->save($user);
         
         return response()->json([
@@ -63,10 +105,11 @@ class GroupController extends Controller
     }
     public function userGroups(Request $request){
         
-        $user = User::find($request->user_id);
+        $user = User::find(auth()->user()->id);
         
         return response()->json([
             'user_groups' => $user ->groups
         ],200);
+    
     }
 }
